requirements.txt
networkx
numpy
sentence-transformers
ollama
requests
ddgs
arxiv
apscheduler
flask
flask-socketio
python-dotenv
pydantic
loguru
pytest
faiss-cpu
